+++ NEWS.md
@@ -1,3 +1,9 @@
+# brazilmaps 0.2.0
+
+* Added Intermediary and Imediate regions
+* "geo.filter" argument in get_brmap changed to "geo_filter"
+* "class" argument in get_brmap changed to "as"
+
 # brazilmaps 0.1.0
 
 * Initial version.